sort(), similar() add
